kimi_k3: PP review round 2 -- the core hook without a docstring, the pipelining entry back in parallelize.py, comments to one line
pipeline_llm keeps only the stage_class parameter (no docstring, the file
has none) and the pyrefly suppression the hook had dropped; the pipelining
entry, the split and the stage lookup live in parallelize.py like every
other model's, layout.py and pipeline_stage.py stay; the full-attention
helper, the debug registry comment and the residual docstring are one line
or upstream's own.
